requirements.txt
aiomysql==0.2.0
aiosmtplib==3.0.2
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
ariadne==0.26.1
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
autopep8==2.3.2
bcrypt==4.2.0
billiard==4.2.1
black==25.1.0
blinker==1.9.0
celery==5.4.0
celery-redbeat==2.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastapi-mail==1.4.2
flower==2.0.1
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
humanize==4.12.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
mypy-extensions==1.0.0
numpy==2.2.0
opencv-python==4.10.0.84
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic-extra-types==2.10.2
pydantic-settings==2.7.0
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.11.1
PyMySQL==1.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2025.1
PyYAML==6.0.2
redis==4.6.0
rich==13.9.4
rich-toolkit==0.13.2
ruff==0.7.4
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
strawberry-graphql==0.260.2
tenacity==9.0.0
threadpoolctl==3.5.0
time-machine==2.16.0
tomli==2.1.0
tornado==6.4.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yapf==0.43.0
